ref: Expand spans v2 before inbound filters
